+++ src/settings.js
@@ -1,3 +1,3 @@
 export const PLATFORM_NAME = 'WhiskerTing';
 export const PLUGIN_NAME = 'homebridge-whisker-ting';
-export const PLUGIN_VERSION = '1.0.7';
+export const PLUGIN_VERSION = '1.0.8';